Implement game functions  into "game()"
Added new variables for the global scores and the round counter.

The score is now showed in the returned string of playRound().

Added two new functions: showScore() that returns the global score and
gameResults() that determine the winner of the game and returns a string
with the global score and the winner of the game.

lastly, added a function called "game" to implement all the necessary
functions for the game logic.