Add UI necessary assets and CSS file
To the assets directory were added the following folders: css, design,
fonts and images. All previous folders with its respective files.

To the styles.css, were added : a CSS RESET, custom properties on the
:root pseudo-class and font faces for the font family that will be
used in the page.

To the index.html, all the inline style that was previous used, was
removed and instead a style sheet link to the styles.css file was added.

Lastly, to the README.md, information about the UI were added.