Add header section
Added the header section to the html with all its respective content,
this section contains the logo and game score.

Added the corresponding css styles to the header section and also some
general styling.

Changed the css color variables from hsl values to hex values, this to
have easy to read values and to use the same value type in all the
colors.

Lastly, added a new css variable for the border radius property that
will be used in the project.