Add CC container and respective JS algorithm
Added the CC (Choices Comparison) container with its respective content
to display the player and computer choices and also the round result.

Added necessary styles to game-choice-section to make it grow in and
take the free space in the main container.

Added necessary styles for the choice circles in the game comparison
section.

Added necessary styles for the presentation of the player and computer
choices and the game results.

Added the necessary JS functions to have a transition between the game
choice container and the choice comparison container, selecting the
necessary elements from the DOM for its utilization in the functions.