Add basic choice comparison functionality
Added a function to get the player choice for its comparison with the
computer choice, also created a function that compares both choices and
give a result message that will be used to display it between both
choices circles.

Adjusted the size and border width of the choice circles within the
choices comparison container.