Add the play again button functionality
Added the functionality to play another game round, this by creating a
function that  hide the choices comparison container and show the game
choices container to once again be able to select an option. Also was
created a function to remove the choices from the last round and leave
the container ready for the next round choices.