Add basic score and round tracking
Two new variables were created, gameScore and roundNumber to track both
players score and the current round of the game.

Added three new functions: updateScore() for update de game score,
getGameFinalResults() to give the result of the match after the 5 rounds
and resetGame() to reset the game score and also the number of the
round.

This variables and functions were implemented to give the user the game
score every round but for the moment its implemented only in the
console. The next step will be to add the game score to the UI.