Add setTimeout function for UI transitions
The setTimeout function was used to have transitions with a time delay
between the player and computer selection and the results of the round/
game.

Added necessary css classes to style the placeholder element for the
computer choice.

The necessary functions for the transition from the game choices
container to the choices comparison container with the game/round
results were created.